+++ projects/journey-maps-client/src/lib/components/teaser/teaser.component.spec.ts
@@ -2,8 +2,8 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 import {TeaserComponent} from './teaser.component';
-import {JourneyMapsClientModule} from '../../../../journey-maps-client.module';
-import {TestDataService} from '../../../../services/test-data.service';
+import {JourneyMapsClientModule} from '../../journey-maps-client.module';
+import {TestDataService} from '../../services/test-data.service';
 import {HttpClientModule} from '@angular/common/http';
 
 describe('TeaserComponent', () => {
@@ -19,13 +19,13 @@ describe('TeaserComponent', () => {
     });
     fixture = TestBed.createComponent(TeaserComponent);
     component = fixture.componentInstance;
-    component.shouldRender = true;
+    component.rendered = true;
     closeClicked = false;
   });
 
   it('should emit closeClicked event if close button is clicked', () => {
     component.closeClicked.subscribe(() => closeClicked = true);
-    component.selectedMarker = testData.createMarker();
+    component.templateContext = testData.createMarker();
     fixture.detectChanges();
     expect(closeClicked).toBeFalse();
     getCloseButtonDe().nativeElement.click();